Added a visualization of solving the 8-puzzle.
The 8-puzzle works quite well; there are play/pause/forward/backward/skip to start/skip to end buttons, a slider for the speed, and one can paste in the solving path for an 8-puzzle from python. Visually, things could be made a little nicer though; in particular, the puzzle itself should be larger and centered.
Updated the navbar to have links between the graph class pages (the spacing needs fixing).
Also added a todo list to keep track of next steps.